Remove automatic tagging of image_name inside manifest.json
Instead; split off start-point/ dirs into their own repos
in their own github organisation. This allows image tags
to be explicit.
Plan is to extend [cyber-dojo start-point create] command
so the GIT_URL can optionally be prefixed with a TAG. Eg
7686e9d@https://github.com/cyber-dojo-start-points/gcc-assert
It should be enought to just look for an @ character in the
8th position.